29 add reactions functionalites for posts and comments

+++ src/main/java/com/social_media_platform/social_media_platform_backend/repositiries/ReactionTypeRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.social_media_platform.social_media_platform_backend.databaseTables.ReactionType;
 
 @Repository
-public interface ReactionTypeRepository extends JpaRepository<ReactionType, Integer> {}
+public interface ReactionTypeRepository extends JpaRepository<ReactionType, Long> {}
